v.13.0.0 - Migrate Functions to vuex mutations
