+++ feedcrawler/web/package.json
@@ -7,14 +7,17 @@
     "preview": "vite preview --port 5050"
   },
   "dependencies": {
+    "@popperjs/core": "^2.11.4",
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.8.1",
     "vue": "^3.2.31",
+    "vue-router": "^4.0.14",
     "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^2.2.2",
+    "@vue/cli-plugin-router": "~5.0.0",
     "vite": "^2.8.4"
   }
 }
