+++ feedcrawler/web/src/components/Lists.vue
@@ -1,6 +1,6 @@
 <script setup>
 import {useStore} from 'vuex'
-import axios from "axios"
+import axios from 'axios'
 
 const store = useStore()
