+++ feedcrawler/web/src/components/__Helper.vue
@@ -1,0 +1,16 @@
+<script setup>
+import "bootstrap/dist/css/bootstrap.min.css"
+import 'bootstrap-icons/font/bootstrap-icons.css'
+import "bootstrap"
+import Head from './_Head.vue'</script>
+
+<template>
+  <main class="text-center">
+    <!--- Main Items -->
+    <Head/>
+  </main>
+</template>
+
+<style>
+@import '.././assets/base.css';
+</style>
