+++ feedcrawler/web/src/main.js
@@ -1,6 +1,7 @@
 import {createApp} from 'vue'
 import {createStore} from 'vuex'
-import axios from "axios"
+import axios from 'axios'
+import router from './router'
 
 import App from './App.vue'
 
@@ -184,4 +185,4 @@ const store = createStore({
     }
 })
 
-createApp(App).use(store).mount('#app')
+createApp(App).use(store).use(router).mount('#app')
